slh-claude-bot/requirements.txt
aiogram==3.13.1
anthropic==0.39.0
python-dotenv==1.0.1
httpx==0.27.2
aiosqlite==0.20.0
